+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ## [Unreleased]
 ### Added
 - support for Gambio GX4.9 and PHP 8.1
+- uses Shopgate Cart Integration SDK 2.10.3
 
 ## [2.9.58] - 2023-04-12
 ### Added